# Conflicts:
#	src/main/java/com/sparta/quizdemo/QuizdemoApplication.java
#	src/main/java/com/sparta/quizdemo/chat/entity/ChatMessage.java
#	src/main/java/com/sparta/quizdemo/chat/entity/ChatRoom.java
#	src/main/java/com/sparta/quizdemo/common/config/WebSecurityConfig.java
#	src/main/java/com/sparta/quizdemo/order/entity/Order.java
#	src/main/java/com/sparta/quizdemo/order/repository/OrderRepository.java
#	src/main/java/com/sparta/quizdemo/order/service/OrderService.java
#	src/main/java/com/sparta/quizdemo/user/entity/User.java

+++ src/main/java/com/sparta/quizdemo/order/service/OrderService.java
@@ -10,7 +10,7 @@ import com.sparta.quizdemo.cart.service.CartServiceImpl;
 import com.sparta.quizdemo.chat.entity.ChatMessage;
 import com.sparta.quizdemo.chat.entity.ChatRoom;
 import com.sparta.quizdemo.common.dto.ApiResponseDto;
-import com.sparta.quizdemo.common.entity.User;
+import com.sparta.quizdemo.user.entity.User;
 import com.sparta.quizdemo.order.dto.OrderRequestDto;
 import com.sparta.quizdemo.order.dto.OrderResponseDto;
 import com.sparta.quizdemo.order.entity.Order;
@@ -22,6 +22,7 @@ import com.sparta.quizdemo.sse.service.NotificationService;
 import com.sparta.quizdemo.user.UserRepository;
 import com.sparta.quizdemo.user.UserRoleEnum;
 import com.sparta.quizdemo.user.UserService;
+import com.sparta.quizdemo.common.entity.UserRoleEnum;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.boot.json.JsonParseException;
@@ -40,9 +41,9 @@ import java.net.URLEncoder;
 import java.time.LocalDateTime;
 import java.util.*;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
-@Slf4j
 public class OrderService {
     private final OrderRepository orderRepository;
     private final OrderItemRepository orderItemRepository;
@@ -85,6 +86,7 @@ public class OrderService {
         lat2 = Double.parseDouble(xyMap.get("y")); // 위도
         lon2 = Double.parseDouble(xyMap.get("x")); // 경도
         city = xyMap.get("z").substring(1);
+        log.info(xyMap.get("z"));
 
         switch (city) {
             case "서울": distance = distance(37.330689, 126.593066, lat2, lon2);
@@ -117,12 +119,17 @@ public class OrderService {
 
         if (orderRequestDto.getPayment().equals(totalPrice)) {
             // 현재 유저의 order 생성
-            Order order = new Order(user, totalPrice, completeTime, orderRequestDto.getRequest());
+            Order order = new Order(user, totalPrice, completeTime, orderRequestDto.getRequest(), orderRequestDto.getOrderComplete());
             orderRepository.save(order);
+
+            List<Order> userOrderList = orderRepository.findAllByUserIdOrderByCreatedAtAsc(user.getId());
+            if (userOrderList.size() > 10) {
+                orderRepository.delete(userOrderList.remove(0));
+            }
 
             if (cartItemList != null) {
                 for (CartItem cartItem : cartItemList) {
-                    OrderItem orderItem = new OrderItem(cartItem, order);
+                    OrderItem orderItem = new OrderItem(cartItem, order, cartItem.getOptionList());
                     orderItemRepository.save(orderItem);
                     cartItemRepository.delete(cartItem);
                 }
@@ -146,22 +153,40 @@ public class OrderService {
         }
     }
 
-    public ResponseEntity<List<OrderResponseDto>> getOrderList() {
+    public ResponseEntity<List<OrderResponseDto>> getDoneOrderList() {
+        List<Order> orderList = orderRepository.findAllByOrderByCreatedAtDesc();
+        if (orderList.isEmpty()) {
+            throw new NullPointerException("주문이 없습니다.");
+        } else {
+            List<OrderResponseDto> orderResponseDtoList = new ArrayList<>();
+            for (Order order : orderList) {
+                if (order.getOrderComplete()) {
+                    OrderResponseDto orderResponseDto = new OrderResponseDto(order);
+                    orderResponseDtoList.add(orderResponseDto);
+                }
+            }
+            return ResponseEntity.status(HttpStatus.OK).body(orderResponseDtoList);
+        }
+    }
+
+    public ResponseEntity<List<OrderResponseDto>> getReadyOrderList() {
         List<Order> orderList = orderRepository.findAll();
         if (orderList.isEmpty()) {
             throw new NullPointerException("주문이 없습니다.");
         } else {
             List<OrderResponseDto> orderResponseDtoList = new ArrayList<>();
             for (Order order : orderList) {
-                OrderResponseDto orderResponseDto = new OrderResponseDto(order);
-                orderResponseDtoList.add(orderResponseDto);
+                if (!order.getOrderComplete()) {
+                    OrderResponseDto orderResponseDto = new OrderResponseDto(order);
+                    orderResponseDtoList.add(orderResponseDto);
+                }
             }
             return ResponseEntity.status(HttpStatus.OK).body(orderResponseDtoList);
         }
     }
 
     public ResponseEntity<List<OrderResponseDto>> getMyOrders(User user) {
-        List<Order> orderList = orderRepository.findAllByUserId(user.getId());
+        List<Order> orderList = orderRepository.findAllByUserIdOrderByCreatedAtDesc(user.getId());
         if (orderList.isEmpty()) {
             throw new NullPointerException("주문이 없습니다.");
         } else {
@@ -177,10 +202,14 @@ public class OrderService {
     public ResponseEntity<ApiResponseDto> cancelOrder(Long orderNo, User user) {
         Order order = orderRepository.findById(orderNo).orElseThrow(() -> new NullPointerException("존재하지 않는 주문 번호입니다."));
         if (user.getId().equals(order.getUser().getId()) || user.getRole().equals(UserRoleEnum.ADMIN)) {
-            orderRepository.delete(order);
-            return ResponseEntity.status(HttpStatus.OK).body(new ApiResponseDto("주문이 취소 되었습니다.", HttpStatus.OK.value()));
+            if (order.getOrderComplete()) {
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new ApiResponseDto("이미 완료된 주문입니다.", HttpStatus.BAD_REQUEST.value()));
+            } else {
+                orderRepository.delete(order);
+                return ResponseEntity.status(HttpStatus.OK).body(new ApiResponseDto("주문이 취소 되었습니다.", HttpStatus.OK.value()));
+            }
         } else {
-            return ResponseEntity.status(HttpStatus.BAD_GATEWAY).body(new ApiResponseDto("해당 주문에 대한 권한이 없습니다.", HttpStatus.BAD_GATEWAY.value()));
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(new ApiResponseDto("해당 주문에 대한 권한이 없습니다.", HttpStatus.BAD_REQUEST.value()));
         }
     }
 
@@ -195,7 +224,20 @@ public class OrderService {
                         tempOrderCount = tempOrderCount + orderItem.getQuantity();
                         orderItem.getProduct().setOrderCount(tempOrderCount);
                     }
-                    orderRepository.delete(order);
+                    Long tempOrderCount = order.getUser().getOrderCount();
+                    order.getUser().setOrderCount(tempOrderCount + 1);
+                    order.setOrderComplete(true);
+
+                    List<Order> totalOrderList = orderRepository.findAllByOrderByCreatedAtAsc();
+                    List<Order> completedOrderList = new ArrayList<>();
+                    for (Order order2 : totalOrderList) {
+                        if (order2.getOrderComplete()) {
+                            completedOrderList.add(order2);
+                            if (completedOrderList.size() > 100) {
+                                orderRepository.delete(completedOrderList.remove(0));
+                            }
+                        }
+                    }
 
                     String url = "";
                     String orderUsername = order.getUser().getUsername();
